WEB-4460 move CATEGORY enum to avoid circular import
